Add css elements to make recipe pages appealing
Add margin and padding to to make recipe pages more visually appealing
and easier to read. Dish title, dish image, and dish introduction were
centered as well as the ingredients and steps containers.

For readability, ingredients and steps container background colors
were changed and for aesthetic purposes, the ingredients and steps
edges were rounded as well as their respective titles.