Add css elements to homepage for visualization
Add images to the recipe list to allow users to have a visual about the
recipe they want to view. The images with their corresponding title
were made into flex items and the list container turned into a flexbox
container. This allowed the dishes to be easily lined up horizontally.

For better user experience, when the mouse hovers over the image or
title, the cursor changes into a pointerand the background color
changes.

Additionally, the link color was changed to be black at all times
(even after visited) and the underline decoration was removed for
aesthetic purposes.